Added unit tests for the functions for functions fs_views, and fs_design_docs. Also, rewrote most of the existing tests to be much simpler, and use only the most essential input components (hash values).
Added tests:
A Javascript View class with views in the local filesystem should raise an exception when an invalid javascript file path is provided for a local view.
A Javascript View class with views in the local filesystem should raise an exception when an empty hash is provided as a container for the local javascript.
A Javascript View class with views in the local filesystem should raise an exception if a passed javascript does not contain a map or reduce tag.
A Javascript View class with views in the local filesystem should create a proper hash containing the information from a single javascript (view) file.
A Javascript View class with views in the local filesystem should properly accrue the views for a given design document/database(without any design docs).
A Javascript View class with design documents in the local filesystem that need to be aggregated should get a hash of design docs under a specific database with multiple design folders.