Added a minor fix which allows local deleted views to also be deleted on the server.
A deleted local view was not being deleted on couch because the loop checking for differences was iterating over the fs views, and comparing them to the db views. The deleted view did not show up in the loop, so there was nothing to compare to the database version. As a result, no differences were detected. Added size check to ensure that the number of fs views matched the number of db views before concluding that they two sets were the same.