Changed couch_tomato's diff style to match git's. The look is a little more intuitive right now. Fixed some dormant bugs with the diff system:
When diffing with a new database in the filesystem, we were getting a crash because we were attempting to pull the non-existant database from couch for comparision. Fixed the issue by adding an is_db? method to check and see if the database actually exists. If not, an empty hash is used for comparision.