Added a CouchTomato config object which will contain information regarding the couchdb address, server port, and basename and added a minor fix to the fs_design_docs method to deal with custom design documents.
When set_config_yml is called, if a path is specified, the method will look at the given yml file and extract the address, port and basename; otherwise, it will detect to see if the environment is rails and attempt to load the config file from Rails.root/config/couch_tomato.yml. If the environment is not rails, or a config file is not found, the fields will be left blank, and it is the users responsibility for setting the values of the object within their rails project. The CouchTomato::Config class should be instantiated as CouchConfig = CouchTomato::Config in an initializer in the users rails project.