Couch tomato is no longer dependent on a rails environment.
In order to use CouchTomato, a Thorfile is required. Once the Thorfile is read, the thor tasks can be used to set up the remaining configuration i.e thor ct:init. All configurations are stored within CouchTomato::Config. If in a rails enviornment, the suffix appended to the database names is the enviornment name. In a standalone viewpoint, the suffixes are not added unless a suffix is specifically added to the generted couch_tomato.yml as couchdb_suffix.